fixed work calculation on testnet
