+++ network/src/consensus.rs
@@ -47,6 +47,8 @@ pub struct ConsensusParams {
 	pub pow_max_adjust_up: u32,
 	/// Optimal blocks interval (in seconds).
 	pub pow_target_spacing: u32,
+	/// Allow minimal difficulty after block at given height.
+	pub pow_allow_min_difficulty_after_height: Option<u32>,
 
 	/// 'Slow start' interval parameter.
 	///
@@ -161,6 +163,7 @@ impl ConsensusParams {
 				pow_max_adjust_down: 32,
 				pow_max_adjust_up: 16,
 				pow_target_spacing: (2.5 * 60.0) as u32,
+				pow_allow_min_difficulty_after_height: None,
 
 				subsidy_slow_start_interval: 20_000,
 				subsidy_halving_interval: 840_000,
@@ -239,6 +242,7 @@ impl ConsensusParams {
 				pow_max_adjust_down: 32,
 				pow_max_adjust_up: 16,
 				pow_target_spacing: (2.5 * 60.0) as u32,
+				pow_allow_min_difficulty_after_height: Some(299187),
 
 				subsidy_slow_start_interval: 20_000,
 				subsidy_halving_interval: 840_000,
@@ -317,6 +321,7 @@ impl ConsensusParams {
 				pow_max_adjust_down: 0,
 				pow_max_adjust_up: 0,
 				pow_target_spacing: (2.5 * 60.0) as u32,
+				pow_allow_min_difficulty_after_height: Some(0),
 
 				subsidy_slow_start_interval: 0,
 				subsidy_halving_interval: 150,
@@ -348,6 +353,7 @@ impl ConsensusParams {
 				pow_max_adjust_down: 0,
 				pow_max_adjust_up: 0,
 				pow_target_spacing: (2.5 * 60.0) as u32,
+				pow_allow_min_difficulty_after_height: Some(0),
 
 				subsidy_slow_start_interval: 0,
 				subsidy_halving_interval: 150,
