+++ verification/src/lib.rs
@@ -88,7 +88,6 @@ mod sapling;
 mod sigops;
 mod timestamp;
 mod work;
-mod work_zcash;
 
 // pre-verification
 mod verify_block;
